Move save_size_hash() as last call in annotate_lithops function
